✨ 제목 20자 이상, 내용 35자 이상 시 ... 축약, 날짜 포맷팅 함수 별도 utils 폴더로 분리
