+++ src/app/data/mockData.tsx
@@ -5,11 +5,12 @@ export const mockData = [
     id: 1,
     member_id: 123,
     nickname: '파즈',
-    title: '양양 서피비치에서 같',
+    title: '양양 서피비치에서 같양양 서피비치에서 같이 서핑하실분',
     start_date: '2024-08-15T00:00:00',
     end_date: '2024-08-16T00:00:00',
     accompany_area: '강원',
-    content: '끝나고 술도 같이 마셔용 20대 남자',
+    content:
+      '끝나고 술도 같이 마셔용 20대 남자 끝나고 술도 같이 마셔용 20대 남자끝나고 술도 같이 마셔용 20대 남자끝나고 술도 같이 마셔용 20대 남자끝나고 술도 같이 마셔용 20대 남자',
     created_at: '2024-07-01T10:00:00',
     profile_image_path: imagePath,
   },
