+++ src/app/utils/textUtils.ts
@@ -1,0 +1,3 @@
+export const truncateText = (text: string, maxLength: number): string => {
+  return text.length > maxLength ? `${text.substring(0, maxLength)}...` : text
+}
